requirements.txt
# runtime
streamlit==1.34.0
requests>=2.32.0
python-dotenv>=1.0.1

# testing
pytest>=8.2.0
pytest-mock>=3.14.0

# linter / formatter
ruff>=0.4.5